Add option to configure CUDA, re-organize imports
